% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertions}
\alias{assertions}
\alias{assert_prod_input_entities}
\alias{assert_user_input_entities}
\title{Assertions}
\usage{
assert_prod_input_entities(x)

assert_user_input_entities(x)
}
\arguments{
\item{x}{`[R object]`

requirements vary.
- `assert_prod_input_entities`: an `integer` vector of valid entity numbers
- `assert_user_input_entities`: an `integer` vector of valid entity numbers}
}
\description{
Assertions specific to NORDCAN metadata.
}
