% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_column_level_spaces}
\alias{get_column_level_spaces}
\alias{get_joint_column_level_space}
\title{Column Level Spaces}
\usage{
get_column_level_spaces(col_nms)

get_joint_column_level_space(col_nms)
}
\arguments{
\item{col_nms}{\verb{[character]} (mandatory, no default)

character string vector of columns names for which to retrieve column
space(s)}
}
\description{
Retrieve allowed levels for categorical NORDCAN columns.
}
\section{Functions}{
\itemize{
\item \code{get_joint_column_level_space}: retrieves a \code{data.table} which determines
allowed combinations of \code{col_nms}
}}

