% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{global_nordcan_settings}
\alias{global_nordcan_settings}
\alias{set_global_nordcan_settings}
\alias{get_global_nordcan_settings}
\title{NORDCAN Settings}
\usage{
set_global_nordcan_settings(
  work_dir,
  stat_cancer_record_count_year_first,
  stat_cancer_record_count_year_last,
  stat_prevalent_subject_count_year_first,
  stat_prevalent_subject_count_year_last,
  stat_survival_follow_up_year_first,
  stat_survival_follow_up_year_last
)

get_global_nordcan_settings()
}
\description{
Set and get settings for NORDCAN software.
}
