% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{in_entity_set}
\alias{in_entity_set}
\title{Entity Detection}
\usage{
in_entity_set(x, entities)
}
\arguments{
\item{x}{`[data.table]` (mandatory, no default)

dataset in which to look for records in specific entities}

\item{entities}{`[integer]` (mandatory, no default)

set of entities to look for}
}
\value{
Returns a logical vector of length `nrow(x)`.
}
\description{
Detect records that have at least one of the supplied entities.
}
