% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_specifications.R
\name{nordcan_column_specifications}
\alias{nordcan_column_specifications}
\alias{nordcan_column_format}
\title{Column Specifications}
\usage{
nordcan_column_specifications(column_name)

nordcan_column_format(column_name)
}
\arguments{
\item{column_name}{\verb{[character]} (mandatory, no default)
name of one column for which to retrieve specifications}
}
\value{
\itemize{
\item \code{nordcan_column_specifications}: a \code{list} of specifications; elements of
the list vary by format,
but they all have element \code{format} (character string)
}

\itemize{
\item \code{nordcan_column_format}: a \code{character} string that names the format; e.g.
\code{"ID"} or \code{"Date"}
}
}
\description{
Retrieve NORDCAN column specifications.
}
