% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{nordcan_entity_levels_by_icd10_and_sex}
\alias{nordcan_entity_levels_by_icd10_and_sex}
\title{NORDCAN Entities}
\format{
\code{data.table} with columns
\itemize{
\item \code{icd10}, of class \code{character}
\item \code{sex}, of class \code{integer}
and the columns for entities by level, each an \code{integer} column.
}
}
\usage{
nordcan_entity_levels_by_icd10_and_sex()
}
\description{
Retrieve definition table of NORDCAN entities by ICD-10 and sex.
}
