% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nordcan_iarccrgtools_tool_names}
\alias{nordcan_iarccrgtools_tool_names}
\title{NORDCAN IARC CRG Tools Tool Names}
\usage{
nordcan_iarccrgtools_tool_names()
}
\description{
Retrieve names of tools of IARC CRG Tools supported by the
NORDCAN R framework.
}
