% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{nordcan_metadata}
\alias{nordcan_metadata}
\alias{nordcan_metadata_icd10_to_entity}
\alias{nordcan_metadata_entity_usage_info}
\alias{nordcan_metadata_icd10_vs_icd7_icd8_icd9}
\alias{nordcan_metadata_entity_by_sex_icd10}
\alias{nordcan_metadata_entity_by_sex}
\alias{nordcan_metadata_entity_no_set}
\title{NORDCAN Metadata}
\usage{
nordcan_metadata_icd10_to_entity()

nordcan_metadata_entity_usage_info()

nordcan_metadata_icd10_vs_icd7_icd8_icd9()

nordcan_metadata_entity_by_sex_icd10()

nordcan_metadata_entity_by_sex()

nordcan_metadata_entity_no_set(entity_no_set_name)
}
\description{
Retrieve definition tables on NORDCAN datasets and their contents.
}
