% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_specifications.R
\name{nordcan_metadata_column_specifications}
\alias{nordcan_metadata_column_specifications}
\alias{nordcan_metadata_column_format}
\title{Column Specifications}
\usage{
nordcan_metadata_column_format(column_name)

nordcan_metadata_column_specifications(column_name)
}
\arguments{
\item{column_name}{`[character]` (mandatory, no default)
name of one column for which to retrieve specifications}
}
\value{
- `nordcan_metadata_column_format`: a `character` string that names the format; e.g.
  `"ID"` or `"Date"`

- `nordcan_metadata_column_specifications`: a `list` of specifications; elements of
  the list vary by format,
  but they all have element `format` (character string)
}
\description{
Retrieve NORDCAN column specifications.
}
