% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{package_data_utils}
\alias{package_data_utils}
\alias{get_exported_dataset_names}
\alias{get_internal_dataset_names}
\alias{get_exported_dataset}
\alias{get_internal_dataset}
\title{Package Data}
\usage{
get_exported_dataset_names()

get_internal_dataset_names()

get_exported_dataset(dataset_name)

get_internal_dataset(dataset_name)
}
\arguments{
\item{dataset_name}{\verb{[character]} (mandatory, no default)

name of dataset to retrieve.}
}
\description{
Utilities to access datasets contained in R packages.
}
