% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{specification_dataset_source}
\alias{specification_dataset_source}
\title{Specification Dataset Sources}
\usage{
specification_dataset_source(dataset_name)
}
\arguments{
\item{dataset_name}{`[character]` (mandatory, no default)
name of a dataset}
}
\description{
Retrieve sources of specification datasets.
}
