% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertions}
\alias{assertions}
\alias{assert_prod_input_entities}
\alias{assert_user_input_entities}
\title{Assertions}
\usage{
assert_prod_input_entities(x)

assert_user_input_entities(x)
}
\arguments{
\item{x}{\verb{[R object]}

requirements vary.
\itemize{
\item \code{assert_prod_input_entities}: an \code{integer} vector of valid entity numbers
\item \code{assert_user_input_entities}: an \code{integer} vector of valid entity numbers
}}
}
\description{
Assertions specific to NORDCAN metadata.
}
