% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{global_nordcan_settings}
\alias{global_nordcan_settings}
\alias{set_global_nordcan_settings}
\alias{get_global_nordcan_settings}
\title{NORDCAN Settings}
\usage{
set_global_nordcan_settings(
  work_dir,
  participant_name,
  stat_cancer_record_count_first_year,
  stat_prevalent_subject_count_first_year,
  stat_cancer_death_count_first_year,
  stat_survival_follow_up_first_year
)

get_global_nordcan_settings()
}
\arguments{
\item{work_dir}{\verb{[character]} (mandatory, no default)

root directory for any and all (temporarily) stored NORDCAN data; if it
does not exist it is attempted to be created}
}
\description{
Set and get settings for NORDCAN software.
}
