% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{global_nordcan_settings}
\alias{global_nordcan_settings}
\alias{set_global_nordcan_settings}
\alias{get_global_nordcan_settings}
\title{NORDCAN Settings}
\usage{
set_global_nordcan_settings(
  work_dir,
  participant_name,
  stat_cancer_record_count_first_year,
  stat_cancer_death_count_first_year,
  regional_data_first_year
)

get_global_nordcan_settings()
}
\arguments{
\item{work_dir}{\verb{[character]} (mandatory, no default)

root directory for any and all (temporarily) stored NORDCAN data; if it
does not exist it is attempted to be created}

\item{participant_name}{\verb{[character]} (mandatory, no default)

name of NORDCAN participant; e.g. "Denmark" or "Sweden"}

\item{stat_cancer_record_count_first_year}{\verb{[integer]} (mandatory, no default)

first year for which to compute the cancer record count statistics;
e.g. \code{1953L}}

\item{stat_cancer_death_count_first_year}{\verb{[integer]} (mandatory, no default)

first year for the cancer death count statistics; e.g. \code{1953L}}

\item{regional_data_first_year}{\verb{[integer]} (mandatory, no default)

first year for regional data; e.g. \code{1953L}}
}
\description{
Set and get settings for NORDCAN software.
}
\details{
Some first years for statistics are set using \code{set_global_nordcan_settings}.
However, the first year for survival is fixed to
\code{nordcancore::nordcan_metadata_nordcan_year() - 29L}. Likewise the first
year for prevalence is fixed to \code{stat_cancer_record_count_first_year + 20L}.
}
