% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{nordcan_metadata}
\alias{nordcan_metadata}
\alias{nordcan_metadata_nordcan_year}
\alias{nordcan_metadata_icd10_to_entity}
\alias{nordcan_metadata_entity_usage_info}
\alias{nordcan_metadata_icd10_vs_icd7_icd8_icd9}
\alias{nordcan_metadata_entity_by_sex_icd10}
\alias{nordcan_metadata_entity_by_sex}
\alias{nordcan_metadata_entity_no_set}
\alias{nordcan_metadata_icd_by_version_to_entity}
\title{NORDCAN Metadata}
\usage{
nordcan_metadata_nordcan_year()

nordcan_metadata_icd10_to_entity()

nordcan_metadata_entity_usage_info()

nordcan_metadata_icd10_vs_icd7_icd8_icd9()

nordcan_metadata_entity_by_sex_icd10()

nordcan_metadata_entity_by_sex()

nordcan_metadata_entity_no_set(entity_no_set_name)

nordcan_metadata_icd_by_version_to_entity()
}
\description{
Retrieve definition tables on NORDCAN datasets and their contents.
}
\details{
\itemize{
\item \code{nordcan_metadata_nordcan_year} just returns the current NORDCAN year
as an integer.
}
}
