% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{object_code_documentation}
\alias{object_code_documentation}
\title{Code Documentation}
\usage{
object_code_documentation(
  text_file_paths,
  regex,
  head = "@details",
  tail = character(0L),
  extract_arg_list = list(),
  grepl_arg_list = list(perl = TRUE)
)
}
\arguments{
\item{text_file_paths}{\verb{[character]} (mandatory, no default)

passed to \link[codedoc:extract_keyed_comment_blocks]{codedoc::extract_keyed_comment_blocks_}}

\item{regex}{\verb{[character]} (mandatory, no default)

regular expression used to identify which comment blocks to retain;
see \link[codedoc:extract_keyed_comment_blocks]{codedoc::extract_keyed_comment_blocks_}; the regex is applied to
retain only those results that pertain to the keys that match the regex}

\item{head}{\verb{[character]} (mandatory, default \code{"@details"})

these lines will be at the head of the output vector}

\item{tail}{\verb{[character]} (mandatory, default \code{"@details"})

these lines will be at the tail of the output vector}

\item{extract_arg_list}{\verb{[list]} (optional, default \code{list()})

additional args passed to \link[codedoc:extract_keyed_comment_blocks]{codedoc::extract_keyed_comment_blocks_}, other
than \code{text_file_paths}}

\item{grepl_arg_list}{\verb{[list]} (optional, default \code{list(perl = TRUE)})

additional args passed to \link[base:grep]{base::grepl}, other than \code{pattern} and \code{x}.}
}
\description{
Using package \strong{codedoc} capabilities, extract documentation blocks from
your source code (using \link[codedoc:extract_keyed_comment_blocks]{codedoc::extract_keyed_comment_blocks_})
to be added to the R documentation for that object.
}
\details{
so you have a roxygen block for your object you want to document. to
include programmatically generated lines into the doc, use the \verb{@eval}
tag like so:

\verb{#' @eval nordcancore::object_code_documentation("R/my_script.R", "my_obj")}
}
